refactor: rename  partition_mapper_mapping as  partition_mapper_config
